feat: add ExerciseDetailDialog component for enhanced exercise details display
